chatapp/requirements.txt
streamlit
streamlit-chat
openai
azure-cosmos